Always call Wait() on the child process
Fixes #5

Previously Wait() was only called if the parent process finished, which
defeats the purpose of acknowledging the zombie state of the child.

Simplify the code flow, always call Wait() (as part of Run()):
- Start the child
- Start the proxy
wait for both of them to complete.

AND, in addition, install a handler that forwards termination signals
to the child process (which in turn will trigger the completion mentioned)
above.