Expose HTTP semantics to functions
HTTP mostly transfers well into a message as the header and entity
directly map over. However, elements of the request like the method,
path and version are lost. Likewise, on the response the status code is
not controllable.

While we don't want to encourage shoehorning applications inside a riff
function, there are times where control over these elements of the http
request/response are important. In particular the response status code
can have significant meaning.

We can map the elements of HTTP that are currently lost into message
headers to expose the fuller semantics to the function. Most functions
should not need to know or care about these headers.

- request method -> `x-http-method` message header
- request path -> `x-http-path` message header
- request query -> `x-http-query` message header
- request version -> `x-http-version` message header
- `x-http-status` message header -> http status code

Resolves #44