Send errors as JSON when the accept type is application/json (#56)
As discussed in projectriff/node-function-invoker/issues/211, we need to be able to alway send JSON back to end-users of riff functions. In the case where an invoker throws an error, the streaming-http-adapter spits out `text/plain`. This is a first pass of the streaming-http-adapter support `application/json` as a first-class content-type.

Signed-off-by: Jesse Brown <jabrown85@gmail.com>